Guard against bad JSON
